Enum rcc::lexer::TokenKind

pub enum TokenKind {

    PLUS,
    MINUS,
    STAR,
    SLASH,
    DSLASH,
    PERCENT,
    DSTAR,
    TILDE,
    AMP,
    DOUBLEAMP,
    PIPE,
    DOUBLEPIPE,
    CARET,
    LSHIFT,
    RSHIFT,
    EQ,
    LT,
    GT,
    GE,
    LE,
    EQEQ,
    NE,
    BANG,
    PLUSEQ,
    MINUSEQ,
    STAREQ,
    SLASHEQ,
    DSLASHEQ,
    PERCENTEQ,
    AMPEQ,
    PIPEEQ,
    CARETEQ,
    LSHIFTEQ,
    RSHIFTEQ,
    QUESTION,
    PTR_OP,
    INC_OP,
    DEC_OP,
    DOT,
    COMMA,
    SEMICOLON,
    COLON,
    LPAREN,
    RPAREN,
    LBRACKET,
    RBRACKET,
    LBRACE,
    RBRACE,
    ELLIPSIS,
    CONSTANT,
    AUTO_KW,
    BREAK_KW,
    CASE_KW,
    CHAR_KW,
    CONST_KW,
    CONTINUE_KW,
    DEFAULT_KW,
    DO_KW,
    DOUBLE_KW,
    ELSE_KW,
    ENUM_KW,
    EXTERN_KW,
    FLOAT_KW,
    FOR_KW,
    GOTO_KW,
    IF_KW,
    INLINE_KW,
    INT_KW,
    LONG_KW,
    REGISTER_KW,
    RESTRICT_KW,
    RETURN_KW,
    SHORT_KW,
    SIGNED_KW,
    SIZEOF_KW,
    STATIC_KW,
    STRUCT_KW,
    SWITCH_KW,
    TYPEDEF_KW,
    UNION_KW,
    UNSIGNED_KW,
    VOID_KW,
    VOLATILE_KW,
    WHILE_KW,
    ALIGNAS_KW,
    ALIGNOF_KW,
    ATOMIC_KW,
    BOOL_KW,
    COMPLEX_KW,
    GENERIC_KW,
    IMAGINARY_KW,
    NORETURN_KW,
    STATIC_ASSERT_KW,
    THREAD_LOCAL_KW,
    IDENTIFIER,
    STRING,
    BYTES,
    COMMENT,
    WHITESPACE,
    NEWLINE,
    UNKNOWN,
    EOF,

}

Variants

PLUS

MINUS

STAR

SLASH

DSLASH

PERCENT

DSTAR

TILDE

AMP

DOUBLEAMP

PIPE

DOUBLEPIPE

CARET

LSHIFT

RSHIFT

EQ

LT

GT

GE

LE

EQEQ

NE

BANG

PLUSEQ

MINUSEQ

STAREQ

SLASHEQ

DSLASHEQ

PERCENTEQ

AMPEQ

PIPEEQ

CARETEQ

LSHIFTEQ

RSHIFTEQ

QUESTION

PTR_OP

INC_OP

DEC_OP

DOT

COMMA

SEMICOLON

COLON

LPAREN

RPAREN

LBRACKET

RBRACKET

LBRACE

RBRACE

ELLIPSIS

CONSTANT

AUTO_KW

BREAK_KW

CASE_KW

CHAR_KW

CONST_KW

CONTINUE_KW

DEFAULT_KW

DO_KW

DOUBLE_KW

ELSE_KW

ENUM_KW

EXTERN_KW

FLOAT_KW

FOR_KW

GOTO_KW

IF_KW

INLINE_KW

INT_KW

LONG_KW

REGISTER_KW

RESTRICT_KW

RETURN_KW

SHORT_KW

SIGNED_KW

SIZEOF_KW

STATIC_KW

STRUCT_KW

SWITCH_KW

TYPEDEF_KW

UNION_KW

UNSIGNED_KW

VOID_KW

VOLATILE_KW

WHILE_KW

ALIGNAS_KW

ALIGNOF_KW

ATOMIC_KW

BOOL_KW

COMPLEX_KW

GENERIC_KW

IMAGINARY_KW

NORETURN_KW

STATIC_ASSERT_KW

THREAD_LOCAL_KW

IDENTIFIER

STRING

BYTES

COMMENT

WHITESPACE

NEWLINE

UNKNOWN

EOF

Implementations

impl TokenKind

pub const fn to_syntax(self) -> SyntaxKind

Convert a given TokenKind to a SyntaxKind. This is used to convert the tokens from the lexer to the tokens used in the parser and the syntax tree.