Struct rcc::token_set::TokenSet

pub struct TokenSet(u128);

A memory-efficient bit-set for managing and querying a collection of SyntaxKinds / TokenKinds.

The TokenSet struct is a space-efficient data structure for working with collections of SyntaxKinds. It is designed for performance and provides many utility APIs for quickly checking membership, union, intersection, and other set operations.

Tuple Fields
0: u128
Implementations

impl TokenSet

pub const EMPTY: TokenSet = _

An empty TokenSet.

pub const fn new(kinds: &[SyntaxKind]) -> TokenSet

Creates a new TokenSet containing the given SyntaxKinds.

Example

use crate::TokenSet;
use crate::SyntaxKind;

pub const fn union(self, other: TokenSet) -> TokenSet

Returns a new TokenSet containing the union of self and other.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set1 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let set2 = TokenSet::new(&[SyntaxKind::Punctuation, SyntaxKind::Keyword]);

let union_set = set1.union(set2);
assert!(union_set.contains(SyntaxKind::Identifier));
assert!(union_set.contains(SyntaxKind::Punctuation));
assert!(union_set.contains(SyntaxKind::Keyword));

pub const fn contains(&self, kind: TokenKind) -> bool

Returns true if the TokenSet contains the specified SyntaxKind.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
assert!(set.contains(SyntaxKind::Identifier));
assert!(!set.contains(SyntaxKind::Punctuation));

pub const fn len(&self) -> usize

pub fn merge(&mut self, kind: TokenKind)

Example

use crate::TokenSet;
use crate::SyntaxKind;

let mut set = TokenSet::new(&[SyntaxKind::Identifier]);
set.merge(SyntaxKind::Keyword);

assert!(set.contains(SyntaxKind::Identifier));
assert!(set.contains(SyntaxKind::Keyword));

pub fn is_empty(&self) -> bool

Returns true if the TokenSet is empty.

Example

use crate::TokenSet;

let empty_set = TokenSet::EMPTY;
assert!(empty_set.is_empty());

pub fn is_subset(&self, other: TokenSet) -> bool

Returns true if self is a subset of other.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set1 = TokenSet::new(&[SyntaxKind::Identifier]);
let set2 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);

assert!(set1.is_subset(set2));
assert!(!set2.is_subset(set1));

pub fn is_disjoint(&self, other: TokenSet) -> bool

Returns true if self and other have no common elements.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set1 = TokenSet::new(&[SyntaxKind::Identifier]);
let set2 = TokenSet::new(&[SyntaxKind::Keyword]);

assert!(set1.is_disjoint(set2));

pub fn intersection(&self, other: TokenSet) -> TokenSet

Returns a new TokenSet containing the intersection of self and other.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set1 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let set2 = TokenSet::new(&[SyntaxKind::Keyword, SyntaxKind::Punctuation]);

let intersection_set = set1.intersection(set2);
assert!(!intersection_set.contains(SyntaxKind::Identifier));
assert!(intersection_set.contains(SyntaxKind::Keyword));
assert!(!intersection_set.contains(SyntaxKind::Punctuation));

pub fn remove(&mut self, other: TokenSet)

Removes the elements of other from self.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let mut set1 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let set2 = TokenSet::new(&[SyntaxKind::Keyword]);

set1.remove(set2);
assert!(set1.contains(SyntaxKind::Identifier));
assert!(!set1.contains(SyntaxKind::Keyword));

pub fn insert(&mut self, other: TokenSet)

Inserts the elements of other into self.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let mut set1 = TokenSet::new(&[SyntaxKind::Identifier]);
let set2 = TokenSet::new(&[SyntaxKind::Keyword]);

set1.insert(set2);
assert!(set1.contains(SyntaxKind::Identifier));
assert!(set1.contains(SyntaxKind::Keyword));

pub fn difference(&self, other: TokenSet) -> TokenSet

Returns a new TokenSet containing the difference of self and other.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set1 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let set2 = TokenSet::new(&[SyntaxKind::Keyword, SyntaxKind::Punctuation]);

let difference_set = set1.difference(set2);
assert!(difference_set.contains(SyntaxKind::Identifier));
assert!(!difference_set.contains(SyntaxKind::Keyword));
assert!(!difference_set.contains(SyntaxKind::Punctuation));

pub fn symmetric_difference(&self, other: TokenSet) -> TokenSet

Returns a new TokenSet containing the symmetric difference of self and other.

The symmetric difference of two sets is the set of elements which are in either of the sets but not in their intersection.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set1 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let set2 = TokenSet::new(&[SyntaxKind::Keyword, SyntaxKind::Punctuation]);

let symmetric_difference_set = set1.symmetric_difference(set2);

assert!(symmetric_difference_set.contains(SyntaxKind::Identifier));
assert!(!symmetric_difference_set.contains(SyntaxKind::Keyword));
assert!(symmetric_difference_set.contains(SyntaxKind::Punctuation));

pub fn toggle(&mut self, other: TokenSet)

Toggles the elements of other in self.

Elements present in other but not in self will be added to self, and elements present in both other and self will be removed from self.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let mut set1 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let set2 = TokenSet::new(&[SyntaxKind::Keyword, SyntaxKind::Punctuation]);

set1.toggle(set2);
assert!(set1.contains(SyntaxKind::Identifier));
assert!(!set1.contains(SyntaxKind::Keyword));
assert!(set1.contains(SyntaxKind::Punctuation));

pub fn toggle_kind(&mut self, kind: TokenKind)

Toggles the presence of the specified SyntaxKind in self.

If kind is present in self, it will be removed. If kind is not present in self, it will be added.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let mut set = TokenSet::new(&[SyntaxKind::Identifier]);

set.toggle_kind(SyntaxKind::Keyword);
assert!(set.contains(SyntaxKind::Identifier));
assert!(set.contains(SyntaxKind::Keyword));

set.toggle_kind(SyntaxKind::Identifier);
assert!(!set.contains(SyntaxKind::Identifier));
assert!(set.contains(SyntaxKind::Keyword));

pub fn is_superset(&self, other: TokenSet) -> bool

Returns true if self is a superset of other.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set1 = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let set2 = TokenSet::new(&[SyntaxKind::Identifier]);

assert!(set1.is_superset(set2));
assert!(!set2.is_superset(set1));

pub fn complement(self) -> TokenSet

Returns the complement of self, which includes all SyntaxKinds that are not in self.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set = TokenSet::new(&[SyntaxKind::Identifier]);
let complement_set = set.complement();

assert!(!complement_set.contains(SyntaxKind::Identifier));
assert!(complement_set.contains(SyntaxKind::Keyword));

pub fn is_singleton(self) -> bool

Returns true if self contains exactly one SyntaxKind.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let singleton_set = TokenSet::singleton(SyntaxKind::Identifier);
assert!(singleton_set.is_singleton());

let set = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
assert!(!set.is_singleton());

pub fn count(self) -> u32

Returns the number of SyntaxKinds in self.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let set = TokenSet::new
(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
assert_eq!(set.count(), 2);

let empty_set = TokenSet::EMPTY;
assert_eq!(empty_set.count(), 0);

pub fn take(&mut self) -> TokenSet

Removes all elements from self and returns a new TokenSet containing the removed elements.

Example

use crate::TokenSet;
use crate::SyntaxKind;

let mut set = TokenSet::new(&[SyntaxKind::Identifier, SyntaxKind::Keyword]);
let taken_set = set.take();

assert!(taken_set.contains(SyntaxKind::Identifier));
assert!(taken_set.contains(SyntaxKind::Keyword));
assert!(set.is_empty());